upd: remove main.h from txdb.cpp
